Change client ctor to detect vbucket usage from server.
The question about whether or not to use vbuckets is really a
funciton of the type of configuration the server will hand out.
There should be no reason to give a URI, bucket name and whether
or not it's vbucket capable.  That should be discernable from the
configuration supplied by the server.

This is believed to be compatible with how moxi detects cache
buckets from membase buckets.

Change-Id: I640fefc37bc863bf5991aa84c6f9fc7e181f565f