Disable optimization by default, remove assert. SPY-88.
This is only a workaround at the moment for SPY-88, but
it was found that the addition of the assert with a
javadoc change and changing OperationState to volatile
would cause failures in test cases like the
testParallelSetGet() indicating there is a thread
safety issue.

While trying to find the specific cause, it seems the best
approach for now is to disable optimization.

The assert, while correct, is also being disabled since
it seems to trigger issues with getBulk which seems to
call the complete() more frequently than it should.

This combination of changes seems to be a tested, stable
workaround until a proper fix may be implemented.

Change-Id: I755b18aa487247a445657726ae2dc8dad6a27991