WIP: Verify connections are sound before completing them entirely.
Without this change, the client considers the connection to be
good any time a connection has been established.  This is not the
case in some critical situations, such as a hung process or a very
busy remote server.

Solution is to ping the server with a noop, since that can happen
before auth, and use that as a way to verify the thing is actually
alive.

Change-Id: Ie3c0bc967b0705df6e58bf8ef81b158db8576bf3